+++ utils/types/index.ts
@@ -3,3 +3,4 @@ export * from 'utils/types/requests'
 export * from 'utils/types/responses'
 export * from 'utils/types/api'
 export * from 'utils/types/forms'
+export * from 'utils/types/zod'
